Alias to show the middle x lines of a file in an efficient manner.
an alias called 'middle' that functions like 'head' and 'tail',
but shows the middle lines of a file.  Give it a starting line, number
of lines to print, and the filename to print lines from, and it will
print only the lines you specify, using an awk expression to make it
fairly efficient.